requirements.txt
pydrive
pystray
pillow
pyperclip
python-dotenv